PMM-11991 Added logs for API requests
