+++ cli-tests/helpers/containers.ts
@@ -32,3 +32,21 @@ export const cleanUpDocker = async (): Promise<void> => {
   await Promise.all(cmds.map((cmd) => cli.execute(cmd)))
     .catch((err) => console.error(err));
 };
+
+/**
+ * Removes code duplication in "pmm-cli/server/upgrade.spec.ts"
+ * Runs pmm-server using docker command with specified parameters.
+ *
+ * @param   httpPort        binds port 80 redirect
+ * @param   httpsPort       binds port 443 redirect
+ * @param   volumeName      name for the docker volume
+ * @param   containerName   name for the container
+ */
+export const runOldPmmServer = async (httpPort: number, httpsPort: number, volumeName: string, containerName: string) => {
+  await (await cli.exec(`docker run --detach --restart always
+        --publish ${httpPort}:80 
+        --publish ${httpsPort}:443 
+        -v ${volumeName}:/srv 
+        --name ${containerName} 
+        ${process.env.old_server_image}`)).assertSuccess();
+};
